Allow for tagging definitions.
Definitions may now include a `tags` field when created, which is a list
of tag names.

The `JXT.tagged(tagName)` function will return the set of matching
definitions.

The intent is to allow for extensions that are a union of possible
definitions. For example, a Jingle Content defintion might have a
`description` field that could be an RTP description, a FileTransfer
description, etc.